requirements.txt
Flask>=2.2.5
Flask-JWT-Extended>=4.4.4
PyJWT>=2.0.0
fido2>=1.1.0